requirements.txt
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.5
paramiko==3.3.1
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
tk==0.1.0
